+++ src/components/icons/index.ts
@@ -1,6 +1,7 @@
 export { AnalyticsIcon } from "./AnalyticsIcon";
 export { ArchiveIcon } from "./ArchiveIcon";
 export { BoardViewIcon } from "./BoardViewIcon";
+export { CalendarIcon } from "./CalendarIcon";
 export { CardDragIcon } from "./CardDragIcon";
 export { CloseIcon } from "./CloseIcon";
 export { CopyIcon } from "./CopyIcon";